Add files via upload
We implement the Gale - Shapley algorithm to solve the stable matching problem. In this problem, we are considering hospitals and residents. The hospitals have a preference list as to which students they want and the students have a preference list as to which hospital they want. The algorithm uses the Gale-Shapley algorithm to solve this problem. It makes sure that it runs in linear time.